# docs/api.yml
openapi: "3.0.3"
info:
  title: API money manager
  version: "1.0"
servers:
  - url: https://api.server.test
...
paths:
  /user:
    post:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/login_success"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors"
    get:
      security:
        - bearerAuth: []
      tags:
        - user
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user_response"
        "404":
          description: user not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors"
    put:
      security:
        - bearerAuth: []
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user_response"
        "404":
          description: user not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors"
    delete:
      security:
        - bearerAuth: []
      tags:
        - user    
      responses:
        "200":
          description: user delete
        "400":
          description: invalid id
        "404":
          description: user not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors"
  /user/login:
    post:
      tags:
        - user
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/body_login"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/login_success"
        "400":
          description: incorrect credentials
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/errors"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    body_login:
      type: object
      properties:
        username:
          type: string
          example: user
        password:
          type: string
          example: 1234
    login_success:
      type: object
      properties:
        token_type:
          type: string
          example: Bearer
        access_token:
          type: string
          example: token
        expires_in:
          type: integer
          example: 18000
    errors:
      type: object
      properties:
        code:
          type: string
          example: error
        message:
          type: string
          example: error
    user_response:
      type: object
      properties:
        username:
          type: string
          example: username
        email:
          type: string
          example: email@gmail.com
    user:
      type: object
      properties:
        username:
          type: string
          example: username
        password:
          type: string
          example: password
        email:
          type: string
          example: email@gmail.com
...
security:
  - bearerAuth : []